refactored to share functions
